Read deployment values from Vault

+++ cf/push.sh
@@ -9,28 +9,32 @@
 # pip install --upgrade awscli
 # aws --version
 ####################################################################################
-# You can also set these in ~/.bash_profile or by running
-# aws configure (Set region name to be "eu-west-1")
-####################################################################################
-# export AWS_ACCESS_KEY_ID=
-# export AWS_SECRET_ACCESS_KEY=
-####################################################################################
-# You'll need to create a .env.XXXX file for each
-# environment you need to be able to push to, containing
-# the following keys
+# You'll need to be logged into:
 #
-# S3_ACCESS_KEY=
-# S3_SECRET_KEY=
-# S3_BUCKET_PUBLIC=transform-par-beta-production-public
-# S3_BUCKET_PRIVATE=transform-par-beta-production-private
-# S3_BUCKET_ARTIFACTS=transform-par-beta-artifacts
-# PAR_HASH_SALT=
-# PAR_GOVUK_NOTIFY_KEY=
-# PAR_GOVUK_NOTIFY_TEMPLATE=
+#    Gov.uk PaaS
+#       e.g. 
+#         cf login -a api.cloud.service.gov.uk -u chris.moreton@transformuk.com
+#           
+#    Par Beta Vault
+#       vault auth
+#       valut unseal
 ####################################################################################
 
 ENV=$1
 VER=$2
+
+AWS_ACCESS_KEY_ID=`vault read -field=AWS_ACCESS_KEY_ID secret/par/deploy/aws`
+AWS_SECRET_ACCESS_KEY=`vault read -field=AWS_SECRET_ACCESS_KEY secret/par/deploy/aws`
+S3_ACCESS_KEY=`vault read -field=S3_ACCESS_KEY secret/par/env/$ENV`
+S3_SECRET_KEY=`vault read -field=S3_SECRET_KEY secret/par/env/$ENV`
+PAR_HASH_SALT=`vault read -field=PAR_HASH_SALT secret/par/env/$ENV`
+S3_BUCKET_PUBLIC=`vault read -field=S3_BUCKET_PUBLIC secret/par/env/$ENV`
+S3_BUCKET_PRIVATE=`vault read -field=S3_BUCKET_PRIVATE secret/par/env/$ENV`
+S3_BUCKET_ARTIFACTS=`vault read -field=S3_BUCKET_ARTIFACTS secret/par/env/$ENV`
+PAR_GOVUK_NOTIFY_KEY=`vault read -field=PAR_GOVUK_NOTIFY_KEY secret/par/env/$ENV`
+PAR_GOVUK_NOTIFY_TEMPLATE=`vault read -field=PAR_GOVUK_NOTIFY_TEMPLATE secret/par/env/$ENV`
+
+vault seal
 
 # We are in the /cf directory
 
@@ -45,6 +49,11 @@ if [ "$VER" != "" ]; then
     cd build
     
     aws s3 cp s3://transform-par-beta-artifacts/builds/$VER.tar.gz .
+    
+    if [ ! -f $VER.tar.gz ]; then
+       exit
+    fi
+    
     tar -zxvf $VER.tar.gz
     rm $VER.tar.gz
     
@@ -56,15 +65,15 @@ fi
 
 cf push -f manifest.$ENV.yml
 
-cf set-env par-beta-$ENV S3_ACCESS_KEY $S3_ACCESS_KEY
-cf set-env par-beta-$ENV S3_SECRET_KEY $S3_SECRET_KEY
-cf set-env par-beta-$ENV PAR_HASH_SALT $PAR_HASH_SALT
-cf set-env par-beta-$ENV S3_BUCKET_PUBLIC $S3_BUCKET_PUBLIC
-cf set-env par-beta-$ENV S3_BUCKET_PRIVATE $S3_BUCKET_PRIVATE
-cf set-env par-beta-$ENV S3_BUCKET_ARTIFACTS $S3_BUCKET_ARTIFACTS
-cf set-env par-beta-$ENV APP_ENV $ENV
-cf set-env par-beta-$ENV PAR_GOVUK_NOTIFY_KEY $PAR_GOVUK_NOTIFY_KEY
-cf set-env par-beta-$ENV PAR_GOVUK_NOTIFY_TEMPLATE $PAR_GOVUK_NOTIFY_TEMPLATE
+cf set-env par-beta-branch S3_ACCESS_KEY $S3_ACCESS_KEY
+cf set-env par-beta-branch S3_SECRET_KEY $S3_SECRET_KEY
+cf set-env par-beta-branch PAR_HASH_SALT $PAR_HASH_SALT
+cf set-env par-beta-branch S3_BUCKET_PUBLIC $S3_BUCKET_PUBLIC
+cf set-env par-beta-branch S3_BUCKET_PRIVATE $S3_BUCKET_PRIVATE
+cf set-env par-beta-branch S3_BUCKET_ARTIFACTS $S3_BUCKET_ARTIFACTS
+cf set-env par-beta-branch APP_ENV $ENV
+cf set-env par-beta-branch PAR_GOVUK_NOTIFY_KEY $PAR_GOVUK_NOTIFY_KEY
+cf set-env par-beta-branch PAR_GOVUK_NOTIFY_TEMPLATE $PAR_GOVUK_NOTIFY_TEMPLATE
 
 cf restage par-beta-$ENV
